Header style and Warning on export
Re-styled the header and added it to it's own php file to be consistent
across all pages. Moved the login into the navbar (although it actually
exists in the header.php). Added an error message when the user tries to
export the data on Cartographie page without any facility types
selected.